requirements.txt
streamlit
pandas
plotly
psycopg2
requests
numpy
matplotlib.plotly